Fix stale install-dev target
The dependency `mapping-workbench` is not relevant anymore. Reuse user
installation for the right dependency `mapping-toolchain` and move the
dev installation command to a dependency in a dedicated dev requirements
file.

P.S: There is a dependency conflict between the user and dev
requirements, in that the user requirements pull in a specific GitHub
tag/version of `ted-sws`, whereas the dev dependency is on `ted-sws`
HEAD. Making the user install a Make target dependency (or running one
after another) overcomes this situation, allowing the dev dependency to
override the user one.