Update EF16 test pkg CM to include Base XPath
This will be used in updated TED-SWS pipeline code to detect whether or
not it is dealing with an old or new eForms CM/package. Old can mean
both Standard Forms and eForms, where this will _always_ be populated.
In the new eForms CMs, howevever, we can explicitly set this empty.